Move and improve now-obsolete test
